refactor SQL database tests
Previously, the SQL database was deployed _after_ the testing
application, which is wrong. It used to work, but it no longer
does, so it's time to fix it.

This commit introduces a special Arquillian extension that
observes the `BeforeClass` and `AfterClass` events (that is,
events fired by Arquillian just before/after tests from a single
class are executed) and deploy/undeploy the database at that
point in time.

That requires switching off automatic application deployment
during Arquillian Cube initialization. Instead, application
is deployed using the `@OpenShiftResource` annotation. That
annotation is also implemented using a `Before/AfterClass`
observer, so it's important to make sure that the database
deployment observer runs sooner.

A little bit of code cleanup elsewhere is also included.