avoid using "oc new-app" to deploy SQL databases
Previously, we used `oc new-app` to deploy SQL databases
into OpenShift. This command used to create a deployment config,
but since OpenShift 4.5, it creates a deployment instead.

It would be unnecessarily hard to make the test work both
with deployment configs and deployments, so this commit changes
the code to always create deployment configs.

It would be possible to use `oc new-app --as-deployment-config`,
but previous versions of `oc` would fail with unrecognized parameter
and parsing version strings is always fragile.

So instead, we use `oc create dc` and other commands to create
the correct OpenShift resources manually. It takes more commands
than just one, but not that much more:

- `oc create dc test-db --image=...` to create the deployment config
- `oc label dc test-db ...` to add a label (for undeploying)
- `oc set env dc test-db ...` to add environment variables
- `oc expose dc test-db --port=...` to create a service